# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polo Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List all clubs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Club"
    post:
      operationId: createClub
      tags: [clubs]
      summary: Create a club
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClubRequest"
      responses:
        "201":
          description: Club created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
  /clubs/{slug}:
    get:
      operationId: getClubBySlug
      tags: [clubs]
      summary: Get club by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
  /clubs/{clubId}/update:
    put:
      operationId: updateClub
      tags: [clubs]
      summary: Update club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClubRequest"
      responses:
        "200":
          description: Club updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
  /clubs/{clubId}/follow:
    post:
      operationId: followClub
      tags: [spectator]
      summary: Follow a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /clubs/{clubId}/unfollow:
    post:
      operationId: unfollowClub
      tags: [spectator]
      summary: Unfollow a club
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
      required:
        - email
        - password
        - displayName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserWithRoles"
        token:
          type: string
      required:
        - user
        - token
    UserWithRoles:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [spectator, team_manager, admin]
        clubMemberships:
          type: array
          items:
            $ref: "#/components/schemas/ClubMembershipBrief"
        teamAssignments:
          type: array
          items:
            $ref: "#/components/schemas/TeamAssignmentBrief"
      required:
        - id
        - email
        - displayName
        - role
        - clubMemberships
        - teamAssignments
    ClubMembershipBrief:
      type: object
      properties:
        clubId:
          type: string
          format: uuid
        clubName:
          type: string
        clubSlug:
          type: string
        role:
          type: string
          enum: [owner, manager]
      required:
        - clubId
        - clubName
        - clubSlug
        - role
    TeamAssignmentBrief:
      type: object
      properties:
        assignmentId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        teamName:
          type: string
        tournamentId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
      required:
        - assignmentId
        - teamId
        - teamName
        - status
    Club:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        logo96Url:
          type: string
          nullable: true
        logo40Url:
          type: string
          nullable: true
        logoInitials:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
    ClubDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        logo96Url:
          type: string
          nullable: true
        logo40Url:
          type: string
          nullable: true
        logoInitials:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        fields:
          type: array
          items:
            $ref: "#/components/schemas/Field"
        tournaments:
          type: array
          items:
            $ref: "#/components/schemas/Tournament"
        followerCount:
          type: integer
        isFollowing:
          type: boolean
      required:
        - id
        - name
        - slug
        - fields
        - tournaments
        - followerCount
        - isFollowing
    CreateClubRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        website:
          type: string
        country:
          type: string
        region:
          type: string
      required:
        - name
        - slug
    UpdateClubRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        website:
          type: string
        country:
          type: string
        region:
          type: string
        logoUrl:
          type: string
        logo96Url:
          type: string
        logo40Url:
          type: string
    Field:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        number:
          type: integer
          nullable: true
        lat:
          type: string
          nullable: true
        lng:
          type: string
          nullable: true
        surfaceType:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - id
        - clubId
...
    Tournament:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
        format:
          type: string
          nullable: true
          enum: [round_robin, single_elim, double_elim, group_knockout, swiss]
        handicapLevel:
          type: string
          nullable: true
        startDate:
          type: string
          nullable: true
        endDate:
          type: string
          nullable: true
        finalsDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, in_progress, completed, archived]
        matchDurationMin:
          type: integer
        gapBetweenMin:
          type: integer
        chukkersPerMatch:
          type: integer
        chukkerDurationMinutes:
          type: integer
        hasThirdPlace:
          type: boolean
        isVisitingLeague:
          type: boolean
        scheduleConfig:
          type: object
          nullable: true
        aiRecommendation:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clubId
        - name
        - status
...